+++ src/main/java/com/LetMeDoWith/LetMeDoWith/presentation/task/dto/RetrieveDowithTaskResDto.java
@@ -21,7 +21,7 @@ public record RetrieveDowithTaskResDto(
         @Schema(description = "시작 시각", defaultValue = "11:30:00") LocalTime startTime,
         @Schema(description = "인증 사진", defaultValue = "[\"https:image\"]") List<String> successImageUrls,
         @Schema(description = "루틴 반복 조건") RetrieveDowithTaskResDto.DowithTaskRoutine routineCondition,
-        @Schema(description = "잔소리 개수", defaultValue = "102") int feedBackCount) {
+        @Schema(description = "잔소리 개수", defaultValue = "102") Long feedBackCount) {
     public static RetrieveDowithTaskResDto from(RetrieveDowithTaskResult result) {
         return RetrieveDowithTaskResDto.builder()
                 .id(result.id())
